chore: move example folder from core to sd-jwt-vc
