+++ examples/sd-jwt-vc-example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sdjwt-core-examples",
+  "name": "sdjwt-vc-examples",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -23,7 +23,8 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@sd-jwt/core": "workspace:*",
-    "@sd-jwt/types": "workspace:*"
+    "@sd-jwt/sd-jwt-vc": "workspace:*",
+    "@sd-jwt/types": "workspace:*",
+    "@sd-jwt/crypto-nodejs": "workspace:*"
   }
 }
